[jscoq] Switch to Web Workers.
JsCoq now runs in a worker and communicates with coq-manager using
messages.

We have reshaped the API quite a bit, which is geared towards
convergence with SerAPI, and imported a few ideas from that development.

Main changes are:

- coq-manager now follows a more reactive approach (as suggested by
  @cpitclaudel) The document model has also evolved.

  This implies a big improvement in ease of document handling. In
  particular, we listen to feedback to color states, and we use the new
  `Cancel` API instead of `EditAt`.

- The library manager is shared with serapi, it has been greatly
  simplified.

- The protocol feature a more robust initialization procedure, even if a
  bit heavier on the client side.

Many many hacks have been removed in favor of more principled solutions;
still, we have a long way to go.

This improves #13, helps a bit with #9 (we can kill the web worker),
and (in an non-optimal way yet) fixes #16.